Update casting and scene controls
